Converted server scripts to classes, added folder in folder loading
