+++ CarApp/package.json
@@ -6,7 +6,7 @@
     "howler": "^2.1.1",
     "music-metadata": "^3.5.3",
     "music-metadata-browser": "^0.7.1",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.12.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-router-dom": "^5.0.0",
