--- CarApp/public/Greatest Hits (Icon)
@@ -1,1 +1,0 @@
-/media/jessy/MyPassport/Music/Queen/Greatest Hits (Icon)
