+++ CarServer/index.js
@@ -3,300 +3,51 @@ const DEV = true;
 /*
 Dependencies
 */
-const http = require('http').createServer(handler);
-const url = require('url');
-const fs = require('fs');
-const path = require('path');
-const util = require('util');
-const io = require('socket.io');
-const ss = require('socket.io-stream');
-const audioData = require('music-metadata');
-const slugify = require('slugify');
+const HttpServer = require('./vendors/http_server/http.js').HttpServer;
+const SocketIO = require('./vendors/socket/socket.js').SocketIO;
+const Audio = require('./vendors/audio/audio.js').Audio;
 
 if(!DEV){
   const GPIORotaryEncoder = require('./vendors/gpio_rotary/rotary.js').GPIORotaryEncoder;
 }
 
-/*
-   GLOBALS
-*/
 
-const FILE_AUDIO_LIBRARY = './audio-library.json';
-const AUDIO_LIBRARY = {};
+class CarServer {
 
-const FILE_VIDEOCLIP_LIBRARY = './audio-library.json';
-const VIDEOCLIP_LIBRARY = {};
+  constructor(){
+    this.HTTP = new HttpServer(process.argv[2]);
+    this.HTTP.init();
+    this.SOCKET = new SocketIO(this.HTTP.SERVER).init(this.socketEvents);
+  }
 
-const PUBLIC_APP_DIR = '../CarApp/public';
+  socketEvents = (socket) => {
+    socket.emit('status', true);
 
-if(!DEV){
-  const RE = new GPIORotaryEncoder().EE;
-}
-
-/*
-  HTTP SERVER
-*/
-const port = process.argv[2] || 80;
-
-// maps file extention to MIME types
-const mimeType = {
-  '.ico': 'image/x-icon',
-  '.html':'text/html',
-  '.js':  'text/javascript',
-  '.json':'application/json',
-  '.css': 'text/css',
-  '.png': 'image/png',
-  '.jpg': 'image/jpeg',
-  '.wav': 'audio/wav',
-  '.mp3': 'audio/mpeg',
-  '.svg': 'image/svg+xml',
-  '.pdf': 'application/pdf',
-  '.doc': 'application/msword',
-  '.eot': 'appliaction/vnd.ms-fontobject',
-  '.ttf': 'aplication/font-sfnt'
-};
-
-http.listen(parseInt(port));
-
-function handler(req, res) {
-  console.log(`${req.method} ${req.url}`);
-
-  // parse URL
-  const parsedUrl = url.parse(req.url);
-
-  // extract URL path
-  // Avoid https://en.wikipedia.org/wiki/Directory_traversal_attack
-  // e.g curl --path-as-is http://localhost:9000/../fileInDanger.txt
-  // by limiting the path to current directory only
-  const sanitizePath = path.normalize(parsedUrl.pathname).replace(/^(\.\.[\/\\])+/, '');
-  let pathname = path.join(__dirname, sanitizePath);
-
-  fs.exists(pathname, function (exist) {
-    if(!exist) {
-      // if the file is not found, return 404
-      res.statusCode = 404;
-      res.end(`File ${pathname} not found!`);
-      return;
-    }
-
-    // if is a directory, then look for index.html
-    if (fs.statSync(pathname).isDirectory()) {
-      pathname += '/index.html';
-    }
-
-    // read file from file system
-    fs.readFile(pathname, function(err, data){
-      if(err){
-        res.statusCode = 500;
-        res.end(`Error getting the file: ${err}.`);
-      } else {
-        // based on the URL path, extract the file extention. e.g. .js, .doc, ...
-        const ext = path.parse(pathname).ext;
-        // if the file is found, set Content-type and send data
-        res.setHeader('Content-type', mimeType[ext] || 'text/plain' );
-        res.end(data);
-      }
+    //Library sources
+    socket.on('getSources', () => {
+      Audio.getSources()
+      .then(aSrc => {
+        socket.emit('emitSources', aSrc);
+      })
+      .catch(err => console.log(err));
     });
-  });
-}
-console.log(`Server listening on port ${port}`);
 
-/*
-   SOCKET.IO
-*/
-io.listen(http).on('connection', function (socket) {
-  //Connection
-  // console.log(socket);
-  socket.emit('status', true);
+    socket.on('addSources', (paths) => {
+      Audio.addSources(paths)
+      .then(status => {
+        socket.emit('addedSources', status);
+      })
+      .catch(err => console.log(err));
+    });
 
-  //Library sources
-  socket.on('addSources', (array) => addSources(socket, array));
-  socket.on('removeSources', (array) => removeSources(socket, array));
-  socket.on('getSources', (type) => getSources(socket, type));
+    socket.on('removeSources', (array) => Audio.removeSources(socket, array));
 
-  if(!DEV){
-    RE.on('Press', (bool) => socket.emit('press', bool));
-    RE.on('Rotate', (side) => socket.emit('rotate', side));
-  }
-});
-
-//Socket event - load all files found in sources parameter and write a library of file information
-function addSources(socket, aSrc){
-   if(typeof aSrc === 'string') aSrc = [aSrc];
-
-   const songList = [];
-   const folderItems = getItemsFromDirs(aSrc);
-   const fileQuery = getSongsFromItems(folderItems);
-   linkFoldersToApp(aSrc);
-
-   fileQuery.then(songs => {
-      for(let song of songs){
-         songList.push(song);
-      }
-
-      AUDIO_LIBRARY.files = songList;
-      writeFile(FILE_AUDIO_LIBRARY, JSON.stringify(AUDIO_LIBRARY));
-      socket.emit('addedSources', true);
-   });
-}
-
-// TODO
-function removeSources(socket, aSrc){
-   socket.emit('removedSources', true);
-}
-
-//Returns json library
-function getSources(socket, type){
-   let file;
-   switch(type){
-      case 'music':
-         file = FILE_AUDIO_LIBRARY;
-         break;
-      case 'video_clip':
-         file = FILE_VIDEOCLIP_LIBRARY;
-         break;
-      default:
-         console.log('Unregistered Source Type');
-         socket.emit('emitSources', false);
-         return;
-   }
-
-   const libraryRaw = fs.readFileSync(file);
-   const aSrc = JSON.parse(libraryRaw).files;
-
-   socket.emit('emitSources', aSrc);
-}
-
-/*
-   Files
-*/
-
-//create symlink from file folder to public app dir
-function linkFoldersToApp(aSrc){
-  for(const src of aSrc){
-    const path = src.split('/');
-    const folder = path[path.length - 1];
-    const publicDir = PUBLIC_APP_DIR + '/' + folder;
-
-    if(!fs.existsSync(publicDir)){
-      fs.symlinkSync(src, publicDir);
+    if(!DEV){
+      RE.on('Press', (bool) => socket.emit('press', bool));
+      RE.on('Rotate', (side) => socket.emit('rotate', side));
     }
   }
-  AUDIO_LIBRARY.paths = aSrc;
-  writeFile(FILE_AUDIO_LIBRARY, JSON.stringify(AUDIO_LIBRARY));
+
 }
 
-//get list of files from source directories
-function getItemsFromDirs(aSrc){
-   const folderItems = [];
-   for(let dir of aSrc){
-      if(dir[dir.length-1] !== '/') dir += '/';
-      const files = fs.readdirSync(dir);
-      for(let file of files){
-         folderItems.push({
-            path: dir,
-            file: file,
-         });
-      }
-   }
-
-   return folderItems;
-}
-
-//Get file info from song items
-function getSongsFromItems(items){
-   let resolveCount = 0;
-   const songMetaList = [];
-
-   const promise = new Promise((resolve, reject) => {
-      let miscFileCount = 0;
-      for(let file of items){
-         const src = file.path + file.file;
-         if(containsAudioExtension(file.file)){
-            const meta = getSongMeta(src);
-
-            meta.then(data => {
-               resolveCount++;
-               songMetaList.push(getSongObject(file, data));
-               if(resolveCount === items.length - 1 - miscFileCount) resolve(songMetaList);
-            });
-         } else {
-            miscFileCount++;
-         }
-      }
-   });
-
-   return promise;
-}
-
-//Write library to selected file library
-function writeFile(file, data){
-   const wLibraryStrean = fs.createWriteStream(file, 'utf-8');
-
-   wLibraryStrean.write(data);
-}
-
-/*
-   Music
-*/
-
-// Returns true or false if audio file is supported
-function containsAudioExtension(file){
-   const supported = ['mp3', 'flac', 'm4a'];
-
-   for(let extension of supported){
-      if(file.includes('.' + extension)) return true;
-   }
-
-   return false;
-}
-
-//returns audio metadata
-function getSongMeta(file){
-  const meta = audioData.parseFile(file, {native: true});
-
-  return meta;
-}
-
-//returns a formated audio object for audioLibrary
-function getSongObject(file, meta){
-   const folder = file.path.split('/');
-   const clientPath = '/' + folder[folder.length - 2];
-   return {
-      title: (meta.common.title ? meta.common.title : file.file),
-      artist: (meta.common.artist ? meta.common.artist : undefined),
-      album: (meta.common.album ? meta.common.album : undefined),
-      path: clientPath,
-      file: file.file,
-   }
-}
-
-
-//Stingray video DB to myDB
-function convertStingrayToLibrary(){
-   const songlist = [];
-   const stingray = fs.readFileSync('./assetInfo.json');
-   const json = JSON.parse(stingray);
-   const playlist = Object.keys(json).map(key => (json[key]));
-
-   const list = Object.values(playlist);
-   for(let item of playlist){
-      Object.keys(item).map(key => {
-         const song = item[key];
-         console.log(song.title + song.artist + song.albumTitle + song.assetId);
-         if(song.title && song.artist && song.assetId){
-            songlist.push({
-               title: song.title,
-               artist: song.artist,
-               album: song.albumTitle,
-               path: './',
-               file: song.assetId,
-            });
-         }
-      });
-      // break;
-   }
-   console.log(songlist);
-   writeFile('./video.json', JSON.stringify(songlist));
-}
-// convertStingrayToLibrary();
+const carServer = new CarServer();
